gw_crawler/malicious_file_crawler/requirements.txt
attrs==20.2.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.1
cssselect==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
jmespath==0.10.0
lxml==4.5.2
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-magic==0.4.18
queuelib==1.5.0
requests==2.24.0
Scrapy==2.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
Twisted==20.3.0
urllib3==1.25.10
w3lib==1.22.0
wget==3.2
zope.interface==5.1.0
python-dotenv == 0.13.0
boto3        == 1.14.56
botocore     == 1.17.56
pytz==2020.1
minio==6.0.0
python-dotenv == 0.13.0
scrapy-dotpersistence==  0.3.0
scrapyd       ==         1.2.1
scrapyd-client==         1.2.0a1
python-scrapyd-api==2.1.2

